+++ src/app/models/job-board/models.index.ts
@@ -1,0 +1,3 @@
+export {Skill} from './skill';
+export {Company} from './company';
+export {Professional} from './professional';