+++ src/app/models/setting/paginator.ts
@@ -1,17 +1,16 @@
 export interface Paginator {
     current_page: number;
-    first_page_url?: number;
-    from?: number;
-    last_page?: number;
+    first_page_url?: string;
+    from?: string;
+    last_page?: string;
     last_page_url?: string;
     links?: Link[];
     next_page_url?: string;
     path?: string;
     per_page: number;
     prev_page_url?: string;
-    to?: number;
+    to?: string;
     total?: number;
-    system?:number;
 }
 
 export interface Link {