+++ src/app/pages/auth/auth.module.ts
@@ -8,30 +8,30 @@ import {AuthRoutes} from './auth.routing';
 // PrimeNG Modules
 import {PasswordModule} from 'primeng/password';
 import {ButtonModule} from 'primeng/button';
-import {MessagesModule} from 'primeng/messages';
 import {InputTextModule} from 'primeng/inputtext';
-import {MessageModule} from 'primeng/message';
 import {DropdownModule} from 'primeng/dropdown';
-import {ConfirmationService, MessageService} from 'primeng/api';
 import {CaptchaModule} from 'primeng/captcha';
+import {TooltipModule} from 'primeng/tooltip';
+import {AutoCompleteModule} from 'primeng/autocomplete';
+import {DividerModule} from 'primeng/divider';
+import {CheckboxModule} from 'primeng/checkbox';
 
 // My Components
 import {AppNotFoundComponent} from './app.not-found.component';
 import {AppAccessDeniedComponent} from './app.access-denied.component';
-import {TooltipModule} from 'primeng/tooltip';
-import { PasswordResetComponent } from './password-reset/password-reset.component';
-import { PasswordForgotComponent } from './password-forgot/password-forgot.component';
+import {PasswordResetComponent} from './password-reset/password-reset.component';
+import {PasswordForgotComponent} from './password-forgot/password-forgot.component';
 import {AppLoginComponent} from './login/app.login.component';
-import {AutoCompleteModule} from 'primeng/autocomplete';
 import {UserUnlockedComponent} from './user-unlocked/user-unlocked.component';
 import {UnlockUserComponent} from './unlock-user/unlock-user.component';
 import {AppUnderMaintenanceComponent} from './app.under-maintenance.component';
 import {SecurityQuestionComponent} from './security-question/security-question.component';
-import {CheckboxModule} from 'primeng/checkbox';
-import { ChangePasswordComponent } from './login/change-password/change-password.component';
-import { SelectInstitutionRoleComponent } from './login/select-institution-role/select-institution-role.component';
-import {DividerModule} from 'primeng/divider';
+import {ChangePasswordComponent} from './login/change-password/change-password.component';
+import {SelectRoleComponent} from './login/select-institution-role/select-role.component';
 import {RegisterSocialiteUserComponent} from './register-socialite-user/register-socialite-user.component';
+import {AppUnregisteredUserComponent} from './app.unregistered-user.component';
+import {SharedModule} from '../shared/shared.module';
+import {RippleModule} from 'primeng/ripple';
 
 @NgModule({
     imports: [
@@ -42,14 +42,14 @@ import {RegisterSocialiteUserComponent} from './register-socialite-user/register
         InputTextModule,
         PasswordModule,
         ButtonModule,
-        MessagesModule,
-        MessageModule,
         DropdownModule,
         TooltipModule,
         AutoCompleteModule,
         CaptchaModule,
         CheckboxModule,
         DividerModule,
+        SharedModule,
+        RippleModule,
     ],
     declarations: [
         AppNotFoundComponent,
@@ -62,8 +62,9 @@ import {RegisterSocialiteUserComponent} from './register-socialite-user/register
         AppUnderMaintenanceComponent,
         SecurityQuestionComponent,
         ChangePasswordComponent,
-        SelectInstitutionRoleComponent,
-        RegisterSocialiteUserComponent
+        SelectRoleComponent,
+        RegisterSocialiteUserComponent,
+        AppUnregisteredUserComponent,
     ],
     providers: []
 })